pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-ashby-connector"
version = "0.1.0"
description = "Model Context Protocol (MCP) server implementation for Ashby ATS integration"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["mcp", "llm", "ashby", "ats"]
dependencies = [
    "mcp",
    "httpx>=0.27",
    "tenacity>=8.0",
    "python-dotenv",
]
license = { text = "MIT" }
authors = [
    { name = "Nigel Xavier", email = "nigel@disisnotawebsite.com" }
]

[project.urls]
Homepage = "https://github.com/nxrobins/ashby-mcp"

[project.scripts]
ashby-mcp = "src.ashby:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
]
evals = [
    "anthropic>=0.39",
    "pyyaml>=6.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "live: tests that hit the real Ashby API (require ASHBY_API_KEY; skipped otherwise)",
]
testpaths = ["tests"]
pythonpath = ["src"]
